ref(node): Move mkdirpSync to a separate file
